# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for M.B. Convent H. School website
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface NewsItem {
  id: number;
  title: string;
  summary: string;
  /** @nullable */
  content?: string | null;
  date: string;
  category: string;
  /** @nullable */
  imageUrl?: string | null;
}

export interface NewsInput {
  title: string;
  summary: string;
  content?: string;
  date: string;
  category: string;
  imageUrl?: string;
}

export interface FacultyMember {
  id: number;
  name: string;
  subject: string;
  qualification: string;
  experience: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  designation?: string;
}

export interface GalleryImage {
  id: number;
  title: string;
  imageUrl: string;
  category: string;
  /** @nullable */
  description?: string | null;
}

export interface AdmissionInput {
  studentName: string;
  parentName: string;
  email: string;
  phone: string;
  gradeApplying: string;
  message?: string;
}

export interface AdmissionRecord {
  id: number;
  studentName: string;
  parentName: string;
  email: string;
  phone: string;
  gradeApplying: string;
  /** @nullable */
  message?: string | null;
  createdAt: string;
}

export interface ContactInput {
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  subject: string;
  message: string;
}

export interface ContactRecord {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  subject: string;
  message: string;
  createdAt: string;
}

export interface SchoolStats {
  totalStudents: number;
  totalFaculty: number;
  yearsOfExcellence: number;
  passRate: number;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface OpenaiConversationInput {
  title: string;
}

export interface OpenaiMessageInput {
  content: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export interface OpenaiError {
  error: string;
}