# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for M.B. Convent H. School website
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: news
    description: News and events
  - name: faculty
    description: Faculty and staff
  - name: gallery
    description: Gallery images
  - name: admissions
    description: Admissions enquiries
  - name: contact
    description: Contact messages
  - name: openai
    description: OpenAI AI chat operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /news:
    get:
      operationId: listNews
      tags: [news]
      summary: List all news and events
      responses:
        "200":
          description: List of news items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NewsItem"
    post:
      operationId: createNews
      tags: [news]
      summary: Create a news item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsItem"

  /news/{id}:
    get:
      operationId: getNews
      tags: [news]
      summary: Get a news item by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: News item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsItem"
        "404":
          description: Not found

  /faculty:
    get:
      operationId: listFaculty
      tags: [faculty]
      summary: List all faculty members
      responses:
        "200":
          description: List of faculty
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FacultyMember"

  /gallery:
    get:
      operationId: listGallery
      tags: [gallery]
      summary: List gallery images
      responses:
        "200":
          description: List of gallery images
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GalleryImage"

  /admissions:
    post:
      operationId: submitAdmission
      tags: [admissions]
      summary: Submit an admissions enquiry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdmissionInput"
      responses:
        "201":
          description: Enquiry submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdmissionRecord"

  /contact:
    post:
      operationId: submitContact
      tags: [contact]
      summary: Submit a contact message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactRecord"

  /stats:
    get:
      operationId: getStats
      tags: [health]
      summary: Get school summary statistics
      responses:
        "200":
          description: School stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchoolStats"

  /openai/conversations:
    get:
      operationId: listOpenaiConversations
      tags: [openai]
      summary: List all conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiConversation"
    post:
      operationId: createOpenaiConversation
      tags: [openai]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenaiConversationInput"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversation"

  /openai/conversations/{id}:
    get:
      operationId: getOpenaiConversation
      tags: [openai]
      summary: Get conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversationWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
    delete:
      operationId: deleteOpenaiConversation
      tags: [openai]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"

  /openai/conversations/{id}/messages:
    get:
      operationId: listOpenaiMessages
      tags: [openai]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiMessage"
    post:
      operationId: sendOpenaiMessage
      tags: [openai]
      summary: Send a text message and receive a streaming text response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenaiMessageInput"
      responses:
        "200":
          description: SSE stream of assistant text chunks
          content:
            text/event-stream: {}

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    NewsItem:
      type: object
      required: [id, title, summary, date, category]
      properties:
        id:
          type: integer
        title:
          type: string
        summary:
          type: string
        content:
          type: ["string", "null"]
        date:
          type: string
        category:
          type: string
        imageUrl:
          type: ["string", "null"]

    NewsInput:
      type: object
      required: [title, summary, date, category]
      properties:
        title:
          type: string
        summary:
          type: string
        content:
          type: string
        date:
          type: string
        category:
          type: string
        imageUrl:
          type: string

    FacultyMember:
      type: object
      required: [id, name, subject, qualification, experience]
      properties:
        id:
          type: integer
        name:
          type: string
        subject:
          type: string
        qualification:
          type: string
        experience:
          type: string
        bio:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        designation:
          type: string

    GalleryImage:
      type: object
      required: [id, title, imageUrl, category]
      properties:
        id:
          type: integer
        title:
          type: string
        imageUrl:
          type: string
        category:
          type: string
        description:
          type: ["string", "null"]

    AdmissionInput:
      type: object
      required: [studentName, parentName, email, phone, gradeApplying]
      properties:
        studentName:
          type: string
        parentName:
          type: string
        email:
          type: string
        phone:
          type: string
        gradeApplying:
          type: string
        message:
          type: string

    AdmissionRecord:
      type: object
      required: [id, studentName, parentName, email, phone, gradeApplying, createdAt]
      properties:
        id:
          type: integer
        studentName:
          type: string
        parentName:
          type: string
        email:
          type: string
        phone:
          type: string
        gradeApplying:
          type: string
        message:
          type: ["string", "null"]
        createdAt:
          type: string

    ContactInput:
      type: object
      required: [name, email, subject, message]
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        subject:
          type: string
        message:
          type: string

    ContactRecord:
      type: object
      required: [id, name, email, subject, message, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        subject:
          type: string
        message:
          type: string
        createdAt:
          type: string

    SchoolStats:
      type: object
      required: [totalStudents, totalFaculty, yearsOfExcellence, passRate]
      properties:
        totalStudents:
          type: integer
        totalFaculty:
          type: integer
        yearsOfExcellence:
          type: integer
        passRate:
          type: integer

    OpenaiConversation:
      type: object
      required: [id, title, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time

    OpenaiMessage:
      type: object
      required: [id, conversationId, role, content, createdAt]
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time

    OpenaiConversationInput:
      type: object
      required: [title]
      properties:
        title:
          type: string

    OpenaiMessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string

    OpenaiConversationWithMessages:
      type: object
      required: [id, title, createdAt, messages]
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/OpenaiMessage"

    OpenaiError:
      type: object
      required: [error]
      properties:
        error:
          type: string