# lib/api-zod/src/generated/types/openaiMessageInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for M.B. Convent H. School website
 * OpenAPI spec version: 0.1.0
 */

export interface OpenaiMessageInput {
  content: string;
}